+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/DependencyGraphTransformer.kt
@@ -327,8 +327,7 @@ internal class DependencyGraphTransformer(
 
           // Read scopes from annotations
           // We copy scope annotations from parents onto this graph if it's extendable so we only
-          // need
-          // to copy once
+          // need to copy once
           scopes.addAll(graphDeclaration.scopeAnnotations())
 
           includedGraphNodes.putAll(